requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
anthropic>=0.40.0
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
